updated materialdrawer from 4.6 to 5.8
contains a fix for Android 7.0 wherein items would get drawn wrongly aligned when first opening the menu, fixes #40

+++ core/src/main/java/net/mabako/steamgifts/activities/Navbar.java
@@ -118,7 +118,7 @@ public class Navbar {
                         if (fragment instanceof SearchableListFragment)
                             ((SearchableListFragment) fragment).stopSearch();
 
-                        int identifier = drawerItem.getIdentifier();
+                        long identifier = drawerItem.getIdentifier();
                         if (identifier == R.string.login) {
                             activity.requestLogin();
